requirements.txt
Flask==3.0.0
gunicorn==20.1.0
Werkzeug==3.0.1
click>=8.1.3
itsdangerous>=2.1.2
Jinja2>=3.1.2
blinker>=1.6.2
setuptools>=3.0
MarkupSafe>=2.1.1
colorama
uvicorn==0.12.2
accelerate==0.24.1
anyio==3.7.1
certifi==2022.12.7
charset-normalizer==2.1.1
diffusers==0.23.1
filelock==3.9.0
fsspec==2023.10.0
idna==3.4
importlib-metadata==6.8.0
networkx==3.0
numpy==1.24.1
optimum==1.14.1
pandas==2.1.3
Pillow==9.3.0
pip==21.1.2
requests==2.28.1
safetensors==0.4.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
transformers==4.35.2
typing-extensions==4.8.0
urllib3==1.26.13
uvicorn==0.12.2
wheel==0.36.2
yarl==1.9.3
webcolors~=1.13
tomesd~=0.1.3